requirements.txt
asyncpg
django-ninja
fastapi
marshmallow
mypy
orjson
piccolo
psqlpy
psycopg[binary,pool]
sqlalchemy
ruff
tortoise-orm
uvicorn